feat: New command get_db_version.
Added mostly for testing purposes, to force db-initialized event.
Event is emitted by backend when page loads but it requires that cra dev server is running before
Tauri app (so there is frontend to receive event) which is unlikely to happen.
Other solution is to clear beforeDevCommand option in tauri config and start cra server manually.