feat: Independent checkboxes.
All form controls are controlled by Form  component, which supplies controls with values and change handlers.
With this change checkboxes can be independent from the form controller, and have separate value and change handler.
This is usefull for implementing logic into form, for example: showing optional controls, enabling and disabling
some functionality etc.

To use checkbox in independent mode one must pass value and onChange props. The name prop should be empty
or set to value that is ont present in properties of form's initial data.